+++ HBWL-demo/services/service-user/src/main/java/com/hbwl/controller/UserController.java
@@ -1,12 +1,12 @@
 package com.hbwl.controller;
 
+import com.hbwl.common.Result;
 import com.hbwl.pojo.User;
 import com.hbwl.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.UUID;
 
 @RestController
 @RequestMapping("/user")
@@ -16,35 +16,43 @@ public class UserController {
     private UserService userService;
 
     @PostMapping("/register")
-    public String register(@RequestBody User user) {
+    public Result register(@RequestBody User user) {
         int row = userService.addUser(user);
-        if (row == -1) return "参数不能为空";
-        if (row == 0) return "注册用户失败";
-        return "注册用户成功";
+        if (row == -1) return Result.error("参数不能为空");
+        if (row == 0) return Result.error("注册用户失败");
+        return Result.success("注册用户成功");
     }
 
     @DeleteMapping("/{id}")
-    public String deleteUser(@PathVariable("id")UUID id){
+    public Result deleteUser(@PathVariable("id")String id,
+                             @RequestHeader("role") String role){
+        if (!role.equals("admin")) return Result.error("权限不足");
         int row = userService.deleteUserById(id);
-        if (row == -1) return "参数不能为空";
-        if (row == 0) return "删除用户失败";
-        return "删除用户成功";
+        if (row == -1) return Result.error("参数不能为空");
+        if (row == 0) return Result.error("删除用户失败");
+        return Result.success("删除用户成功");
     }
 
     @PutMapping
-    public String updateUser(@RequestBody User user){
+    public Result updateUser(@RequestBody User user,
+                             @RequestHeader("role") String role){
+        if (!role.equals("admin")) return Result.error("权限不足");
         int row = userService.updateUserById(user);
-        if (row == -1) return "参数不能为空";
-        if (row == 0) return "更新用户失败";
-        return "更新用户成功";
+        if (row == -1) return Result.error("参数不能为空");
+        if (row == 0) return Result.error("更新用户失败");
+        return Result.success("更新用户成功");
     }
 
     @GetMapping
-    public List<User> getUsers(@RequestParam(required = false) String id,
-                               @RequestParam(required = false) String username){
+    public Result getUsers(@RequestParam(required = false) String id,
+                               @RequestParam(required = false) String username,
+                               @RequestParam(required = false) String userRole,
+                               @RequestHeader("role") String role){
+        if (!role.equals("admin")) return Result.error("权限不足");
         User user = new User();
         user.setId(id);
         user.setUsername(username);
-        return userService.getUsers(user);
+        user.setRole(userRole);
+        return Result.success(userService.getUsers(user), "查询用户成功");
     }
 }
